Add nova-multi-cell job to experimental queue
Nova has had a multi-cell job (nova-multi-cell) since the
Stein release but is currently non-voting in the check queue
for nova changes. This change adds the job to the experimental
queue for devstack changes so we can test changes to devstack
and make sure they work for the multi-cell job.

Change-Id: Icf31baf6fd4313aec5ecfb9e8f9cbcef1ff7f61d