Create Keystone internal endpoint
Need to create the Keystone internal endpoint to facilitate the
deprecation of the admin endpoint and the transition to the
internal endpoint. With the removal of Keystone v2 APIs, the admin
endpoint is subsequently deprecated because there's no longer a
distinction between admin and internal endpoints.

Notice that on a typical production environment, the controllers
are running on an internal management network, which is segmented
off from the public network per standard security practice.
Therefore, admin endpoint is really functional equivalent to the
internal endpoint, as opposed to the public endpoint.

TODO: we will need to remove the admin endpoint creation once
the transition is completed.

Change-Id: I98b86b0338e859f0a6728200b91dd6295625c08e
Closes-Bug: 1830002