Add NOVA_NOTIFICATION_FORMAT variable
Nova change https://review.opendev.org/603079/ changed the
default configuration to send only unversioned notfications rather
than both versioned and unversioned notifications. This could break
unsuspecting downstream projects (like Watcher) whose CI jobs are
not explicitly configuring nova for the types of notifications
they need but are just relying on getting both per the previous
default of the config option.

This adds a variable which defaults to "both" to match the
nova default (in Stein) but allows downstream CI jobs to easily
configure another value.

Needed by https://review.opendev.org/667161/

NOTE(mriedem): The nova default changed from "both" to
"unversioned" in Train. This stable/stein backport changes the
default in devstack to "both" to match the nova default in Stein.
The goal of the patch is to provide a way to override the config
which this still does and is necessary for the linked Watcher
patch.

Change-Id: Ied9d50b07c368d5c2be658c744f340a8d1ee41e0
(cherry picked from commit b57757ae14a54e78ebd533198564d24af77c51da)