Update patch set 2
Patch Set 2: -Workflow

I still can't get the watcher grenade job to work with this, but I think it's because the setting is being applied by the watcher plugin after nova is upgraded and restarted so it's too late in the process. Might need to land this and backport it and then have the watcher change depend on both so it gets nova configured on the stein side before upgrading to train.

Patch-set: 2
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0