doc: Avoid duplicated entries in toctree
TOC shows duplicated entries from guides.rst now.
Titles in guides.rst and toctree entries in guides.rst
are duplicated. This commit converts titles in guides.rst
to a list to avoid it.

This change also improves the structure of the "Guides" section
in the PDF docs by not using titles in guides.rst and moving
the (hidden) toctree to the bottom of guides.rst.

Change-Id: I2985743116760fc4503d3f63ae83324be356c8d6