WIP: Test using script instead of osc for keystone config
The keystone setup of accounts and roles is taking up about 1/5 of
devstack's runtime. This change uses a small python script to get an
idea for whether or not replacing osc calls with simpler scripts will be
quicker.

The motiviation for this is osc is known to be slow due to pkg_resources
loading. We can check how much of this slowness is actually in that
pkg_resources loading or in our api calls by replacing osc with simple
direct api calls instead.

Chances are if we decide to go further down this route we'll want to
replace whole sections of stack.sh like:
https://opendev.org/openstack/devstack/src/commit/6aeaceb0c4ef078d028fb6605cac2a37444097d8/stack.sh#L1146-L1161
with a single script that can configure everything with the overhead of
one python process startup. That means this change is probably not worth
merging as is even if it functions but should give us dat.

Change-Id: Ifc125b6b33a1c6747b5efd34bb05ebb6f7c6eb49