WIP replace more things with python script
There is a lot of checking that we have to do to keep resources names
and IDs straight when using the SDK. OSC does this for us in the old
code and now we have to do it ourselves. What this means is we only
speed up compared to osc if we remember those values.

Speed up by shifting more of the keystone account creation into a script
so that these values can be remembered which should be faster than OSC.

Change-Id: I112fbeda3b5da23c3a446bd04ade1f35847cdb01