Update patch set 7
Patch Set 7:

(1 comment)

I'd have to take advice from others on what they think devstack is testing at this point; if anything.  Does OSC, or keystone, etc. actually consider this part of the testing of the client and it's interface?  Or is this just busywork on the way getting the cloud ready for real testing and so it doesn't matter how it's done?

Do we just give up that OSC can't be made any faster?  Certainly not a new issue; when the timing stuff was put in I'm sure the changelogs will show OSC client speed was a big part of it.

As mentioned inline, I'd like to see us at least start tracking the timing from the start.  Also a start at getting average numbers from runs to make comparisons.

Patch-set: 7
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0