Update patch set 6
Patch Set 6:

recheck to get some more data points

The current devstack run shows under 7s for the script execution, compared to about 80s that the old sequence of calls took in other runs. While that is a nice improvement, it still less than 10% of the total runtime, so I'm not completely decided yet whether the speedup would be worth the reduction in maintainability and readability.

If we go this path, we probably would need to add tests for the script in order to make sure that it performs all the tasks we expect it to do, because not all of it might get exercised in the jobs that we run.

Patch-set: 6
Reviewer: Gerrit User 13252 <13252@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0