Disable apparmor on SLE systems
The existing code to disable apparmor on SUSE systems only worked for
openSUSE.  On SLE12 (at least), aa-enabled and aa-teardown are not
available.  Even if they were, it is more idiomatic and robust to use
systemd's interface for stop/disable rather than making assumptions
about the underlying implementation.

So replace the references to aa-enabled and aa-teardown with calls to
systemctl.

Change-Id: I8d99c8d743cc1935324e2e4fcb67efaa5241199e