Fix server_listen/server_proxyclient_address for mult-node jobs
Change Iea469128b15298aee61245e702d20603c8d376fb removed the
IPv4-hard-coded VNCSERVER_LISTEN and VNCSERVER_PROXYCLIENT_ADDRESS
values which makes the test_novnc tempest test intermittently fail
when the vnc server_listen and server_proxyclient_address values
just default to 127.0.0.1 and the subnode compute is not on the
same host as the n-novnc service.

This fixes the problem by using the same values as before for
the defaults and also handles being setup to use IPv6.

Change-Id: I0df958e5770de4c391337df4ae90cc2c710ccfa6
Closes-Bug: #1669468