Use OSCaaS to speed up devstack runs
OpenStackClient has a significant amount of startup overhead, which
adds a non-trivial amount of time to each devstack run because it makes
a lot of OSC calls. This change uses the OSC service from [0] to run
a persistent process that handles openstack calls. This removes most
of the startup overhead and in my local testing removes about three
minutes per devstack run.

Currently this is implemented as an opt-in feature. There are likely a
lot of edge cases in projects that use a devstack plugin so turning it
on universally is going to require boiling the ocean. I think getting
this in and enabled for some of the major projects should give us a lot
of the benefit without the enormous effort of making it 100% compatible
across all of OpenStack.

Another thing to note is that the openstack-cli-server doesn't seem to be
able to properly handle image upload commands. The client side has
logic to run "openstack image create" directly instead of calling the
server in that case.

Since the devstack job definitions are used as the base job for many
other jobs, the ansible var is set in the check and gate sections
instead of the job definition itself so they only apply to the jobs
running in devstack itself.

Change-Id: I28e6159944746abe2d320369249b87f1c4b9e24e
0: http://lists.openstack.org/pipermail/openstack-dev/2016-April/092546.html