Fix Redhat to RedHat in apache cfg path selection
ansible_os_family for RHEL is named 'RedHat', cfr this block 
in roles/apache-logs-conf/tasks/main.yaml :

- name: Link apache logs on RedHat
  block:
  - name: Find logs
    find:
      path: "/var/log/httpd"
      file_type: any
    register: redhat_apache_logs
  - name: Dereference files
    stat:
      path: "{{ item.path }}"
    with_items: "{{ redhat_apache_logs.files }}"
    register: redhat_apache_deref_logs
  - name: Create hard links
    file:
      src: "{{ item.stat.lnk_source | default(item.stat.path) }}"
      dest: "{{ stage_dir }}/apache/{{ item.stat.path | basename }}"
      state: hard
    with_items: "{{ redhat_apache_deref_logs.results }}"
    when:
      - item.stat.isreg or item.stat.islnk
  when: ansible_os_family == 'RedHat'

but on line 60 of that same file, it is wrongly referred to as
'Redhat' instead of 'RedHat' (on stable/queens).
The issue is blocking stable/queens CI pipelines on RHEL.

On stable/rocky and above this is fixed.
This patchset addresses this on stable/queens as well.

Change-Id: I8c1a72ec67300a88c85e65af97da4bdf8482f427