Fix worlddump log collection
All credit for figuring this out goes to frickler (and that was the hard
bit so thank you!). The worlddump files were not being collected because
they weren't in our log collection list. Add worlddump to this list
so that we collect these files.

One thing that makes this slightly complicated is the worlddump files
are named with a timestamp and we can't have globs in our collection
list. To address this we create a symlink to the latest version of
the worlddump and collect that name instead (which is static and doesn't
need a glob).

Change-Id: I96ae5f5290546ad25ca434c1106c01354d2d053c