WIP: Configure tempest flavors with a required trait
There was a regression in the libvirt compute driver such
that trying to create a server with any required trait will
fail. This was not caught in nova functional tests since they
use the fake compute driver. This is a simple devstack change
to recreate the bug.

To tickle the bug we have to configure nova-compute with
a custom cpu_model and a cpu_model_extra_flags value.

Depends-On: https://review.opendev.org/681932/

Change-Id: Ic1a94a946012fd2bf31c60f45886626a5196007b
Related-Bug: #1843836