Update patch set 1
Patch Set 1:

Ah, good find. Maybe we can add a retry login, but probably just aborting and deferring to the next run should be good enough.

The next run today didn't propose an update of this one, though, I'm assuming that's because it didn't find any change relative to the current state in master. So we can either abandon this patch or leave it pending until the next real update happens.

Patch-set: 1