Fix six package on opensuse for pip 10
openSUE Leap 15.0 and 15.1 both provide python3-six version 1.11.0.
Since version 1.12.0 was released, pip>=10 recognizes the version
difference and tries to uninstall the distro-provided version and fails.
This change adds another hack to remove the egg-info file for the six
library so that pip can manage it directly. We also have to wait to
install os-testr until after the fixup has happened since trying to
install it triggers the issue.

(WIP: remove depends-on so that this can merge first and fix the pip10
patch)
Depends-on: https://review.opendev.org/561597

Change-Id: I4649abe06b5893a5251bfcdd4234abccde6ceda2