Update patch set 1
Patch Set 1: Code-Review-1

Actually on my second test run I managed to trigger a 500 and it seems like we either need more retries or include a backoff timer like in the stackforge example, maybe both.

http://paste.openstack.org/show/779419/

The good thing is that we seem to be properly aborting now on a 500 instead of treating it like a non-existing plugin.

Patch-set: 1