Update patch set 2
Patch Set 2: Code-Review-1

To me this looks questionable:

a) If nova gets credentials for the cinder user, why do we have different accounts at all? In my understanding, the idea behind this is to keep services separated, if nova gets credentials from other services, the attack surface is increased.

b) If the cinder user doesn't need an admin role, adding it just for nova seems like an unneccessary priviledge increase, too. Also, what about real world deployments? Would you add this step to the Cinder docs? Or in the Nova docs tell installers to change settings for the cinder user?

So I'd like to know why nova can't use the "nova" user for this? And if there is a good reason for that, probably creating a dedicated "nova-cinder" user would seem a better approach to me.

Patch-set: 2
Label: Code-Review=-1