Update patch set 1
Patch Set 1:

> Sounds like there is a fix in Cinder (queens here https://review.opendev.org/#/c/684954/)
 > which could be backported to pike (and maybe ocata) to resolve the
 > issue, so I guess we can see how that goes or just bite the bullet
 > and pin tempest since it's not really safe to be running
 > stable/pike against master tempest when tempest master isn't gating
 > against a pike job for regressions, so we might whack this mole but
 > there will be another one eventually.

I do not think the cinder backport will fix the issue. In my test patch, its v1 version which causing the issue and v1 should not be
returned in GET / as per cinder pike code.
- https://zuul.opendev.org/t/openstack/build/e13e8a408f214e1b9d03b41c23955c7e/log/controller/logs/tempest_log.txt.gz#70152

Something else is causing this issue. 

But pining Tempest is good option to unblock the pike gate.

Patch-set: 1