Configure mysqld with max_allowed_packet=16M
Based on [1] and seeing it's fairly common [2] configure mysql
to use max_allowed_packet=16M to avoid aborted connection issues
seen in grenade runs.

What we're seeing in bug 1844929 is that during a grenade run
upgrading to Train, mysqld is started three times (install, devstack
old side restart, devstack new side restart) and after the final
restart we start losing connections which manifest themselves during
nova scheduling with "Timed out waiting for response from cell"
errors trying to query the nova_cell1 database.

It's not clear what might have changed in Train. To figure that out
we likely need more verbose mysqld logging enabled in a patch that
hits the failure. But for now it seems useful to try and set
max_allowed_packet to a higher size as a workaround.

[1] https://dba.stackexchange.com/questions/19135/mysql-error-reading-communication-packets/19139#19139
[2] http://codesearch.openstack.org/?q=max_allowed_packet&i=nope&files=&repos=

Change-Id: I2310d50dfbd704e522a45e023415c6a58fb43442
Related-Bug: #1844929