Update patch set 3
Patch Set 3:

recheck not sure yet why this job seems to be slower on centos+fedora. we still should be able to get one passing run. we might also consider increasing the job timeout, but 2h are already quite long. the other option might be reducing the number of tests being run, but that task would probably be on tempest.

Patch-set: 3
Label: Verified=0