Handle uninstalled python3
Distros such as Centos 7 do not install python3 with the minimal
install. This ensures that python3 gets installed if it isn't.
Also, python3 package is now python36 when installing from EPEL.
I don't believe anything still uses python3.5 by default.

Ideally, PYTHON3_VERSION would be set *after* ensuring that python3
was installed, but it gets a little tricky with all of the includes.

Change-Id: I7bdfc408b7c18273639ec26eade475856ac43593