Update patch set 1
Patch Set 1:

As I said in the mailing list [1], which was sent *after* this change was approved, I don't see the need to remove this. It's not broken and we have quite a few jobs that still run with postgres. Most are probably in the experimental queue because they can be run on demand to converse gate resources but that doesn't mean pg isn't supported/working/stable.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-October/010198.html

Patch-set: 1