Update patch set 1
Patch Set 1: Code-Review+1

From memory:

At the Pike summit sdague wanted to drop postgres because it wasn't working at the time. In the session on this topic Suse stood up and said they depend on it and will make it work. Since then I don't believe we ever deleted postgres and that was because it was made to work as promised.

I don't know if the recent Suse news played into this at all (mriedem points out that the change to remove this support was pushed in august so probably not), but I think if it is working today then we may as well keep it until it stops working again then bring up removal again in a much broader context (like sdague did).

That said I'm not +2'ing this change because while I have core on devstack I've largely applied it to gate bug fixes (which this sort of now is I guess?) and not kept up with higher level planning.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1