Update patch set 2
Patch Set 2: Code-Review+2 Workflow+1

For some background there was a deprecation message added in:

https://review.opendev.org/#/c/430885/

from sdague's larger discussion and push to remove that wasn't working. There was an attempt to remove that mesage in: https://review.opendev.org/#/c/490083/ which at the time was premature.

Despite apparently being very much in favor of removal based on my comments on the second link I think times have changed, and I'd probably be +2 on 490083 today if it were in a position to merge. In addition I do not agree with how this was done. Personally I was surprised to see this on the ML and there were a bunch of jobs actively broken by this removal and it wasn't clearly communicated. The deprecation message said it would be removed in pike which was so long ago clearly no one actually cared. If we want to go about the removal I'm in agreement with the other comments on this we need to start over again re-add/update the deprecation warning, start a larger discussion, and set a timetable that everyone is aware of.

clarkb: you assumes there is high level planning for devstack. From my PoV it's mostly just been maintenance for the past few years to keep things working.

Patch-set: 2
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1