Pad an extra 1GiB onto tempest flavor root disk
Change I537c299b0cd400982189f35b31df74755422737e in Stein
changed the tempest flavor root disks from 0 to the size
of the image rounded up to the nearest GiB, which for the
small Cirros image we use in the gate is just 1.

Since then, we have seen intermittent ssh issues where the
guest console log shows:

  sh: write error: No space left on device
  ...
  GROWROOT: NOCHANGE: partition 1 is size 2078687. it cannot be grown

This change pads an extra 1GiB of space onto the root disk for
the tempest flavors to avoid running out of space in the guest.

Note that as a result, with the default cirros image we'll have
tempest flavors with a root disk of 2GiB and tempest defaults
volumes it creates to 1GiB, which will fail boot from volume
scenarios so we have to configure tempest's volume_size to be
the max of the root disk of the two flavors used in tempest runs.
Note that we probably should have been doing this anyway, we just
got lucky with the cirros image being rounded up to 1GiB and the
default volume size in tempest being 1GiB.

Change-Id: If19b2bc3523a046162f378cbf1bccf993971b55d
Related-Bug: #1808010