Pad an extra 1GiB onto tempest flavor root disk
Change I537c299b0cd400982189f35b31df74755422737e in Stein
changed the tempest flavor root disks from 0 to the size
of the image rounded up to the nearest GiB, which for the
small Cirros image we use in the gate is just 1.

Since then, we have seen intermittent ssh issues where the
guest console log shows:

  sh: write error: No space left on device
  ...
  GROWROOT: NOCHANGE: partition 1 is size 2078687. it cannot be grown

This change simply pads an extra 1GiB of space onto the root
disk for the tempest flavors to avoid running out of space in
the guest.

Change-Id: If19b2bc3523a046162f378cbf1bccf993971b55d
Related-Bug: #1808010