Update patch set 1
Patch Set 1:

> Hmm, this makes test_image_defined_boot_from_volume fail because we're using a 2gb disk but a 1gb volume. Likely means tempest's volume_size config option needs updating to match that size.

Yeah the default volume_size in tempest.conf is 1:

https://github.com/openstack/tempest/blob/52d7083f76c3b48650a7e6cda5fdad03332b51ed/tempest/config.py#L875

and we're hitting this during boot from volume in the compute API:

https://github.com/openstack/nova/blob/1bfa4626d13d0a73e63745cc4a864ae86d490daf/nova/utils.py#L761

And the flavor has root_gb=2 and the image has min_disk=0:

https://zuul.opendev.org/t/openstack/build/290328c3f6ae455dbb410b421735efc8/log/controller/logs/devstacklog.txt.gz#24340

So it blows up saying the volume is too small.

The root disk size in the flavor is an integer so we can't go like 1.5 there. We'd have to configure tempest's volume_size to be at least the larger of the two flavors used in tempest (which devstack should probably be setting that anyway), but I'm not sure if we'll start having issues creating 2gb volumes in test runs now - if anything it could make runs slightly slower, but maybe not noticeably so. I know we also used to run into issues with how much space was carved out on the host for volumes, but that defaults to 24GB in the gate now (VOLUME_BACKING_FILE_SIZE variable) so it shouldn't be a problem (would we have more than 12 volumes at the same time in a tempest run? probably not).

Patch-set: 1