Update patch set 1
Patch Set 1:

This breaks the centos-7 platform job, which is still running python2 only by default. Might not be a bad idea to drop that job and wait for a working centos-8 job, though.

Also note that the sdk job is broken by novaclient currently, no need to recheck for now.

Patch-set: 1
Label: Verified=0