Update patch set 1
Patch Set 1:

> This breaks the centos-7 platform job, which is still running
 > python2 only by default. Might not be a bad idea to drop that job
 > and wait for a working centos-8 job, though.
 > 
 > Also note that the sdk job is broken by novaclient currently, no
 > need to recheck for now.

Thanks for your advise! 
I found that the devstack's code will execute python installation,if the env does not have python2 installed and python is the default python installation, it will install python2 and set it as default, so executing 'python -c' will be blocked.
Or we can change the python installation to python3, I cannot make sure which way is better to fix the bug.
Please help to follow up on this patch~

Patch-set: 1
Label: Verified=0