Update patch set 5
Patch Set 5:

The main argument isn't really anything about customers--but tons of developers develop for Openstack *on Centos machines*. Devstack not working on Centos is not an option. Every developer running Centos having to manually fix *locally* devstack on Centos is a high support load. Having to go in and write a 6 line patch every now and then is not. Python2 support removal is happening as we speak. There is currently no RDO for Centos 8 (so no OVS packages). So python3 on Centos 7 *has* to work.

Patch-set: 5