Update patch set 5
Patch Set 5:

So far I don't see a follow-on that shows python3 based jobs working on centos7, and I'd suggest we don't merge pieces of support until it is working in the gate.

We have to consider if we as a community want to devote person and compute resources to support a python3 based centos7 job.

It is clear from the stability of the existing job, which I track at [1], people are not that interested in fixing what we have.  For the talk of what "must" or "has" to be supported, there isn't much of a track record of people actually working on the job.  The summary is that we will switch this unstable and under-maintained job to a completely different base python, creating an environment that is realistically not used by even a fraction of a % of production deployments.  IMO this will not be a magnet for further developer time and effort.

Maybe I can be proven wrong with a great groundswell of activity to stabilise the existing centos7 job and convert it.  Personally, I would rather see this effort put towards centos8 reviews, ideally with the input of the work happening for the RDO project [2] and centos cloud SIG such that we can build the devstack-based gate tests in parallel and become/provide some sort of useful component of their ecosystems.

[1] http://people.redhat.com/~iwienand/devstack-status/
[2] https://trunk.rdoproject.org/rhel8-master/deps/latest/

Patch-set: 5