Update patch set 5
Patch Set 5:

> Maybe there would be more willingness to contribute patches for this kind of thing if it wasn't such a headache to merge trivial patches.

It's not about these two lines, as such.  The discussion is much more "should we support a python3 EPEL + centos7 environment" which is a much larger question.

So far I'm not even sure it works; the one run of this I can see failed cryptically trying to make a venv [1].  I think it's a little unfair to say we don't merge things when the patch implies a very different environment, doesn't have testing and the one time it did run it failed.  Feel free to follow-on this patch with testing, Zuul is great like that.

But, for the reasons I expressed, along with RDO having no plans to support this environment [2], I don't think this is going to be time well spent.  Personally I don't plan on spending my limited time dealing with issues with such a job (were it to exist); I'd much rather spend my time working on what we can do with centos8 [3].  My unsolicited advice is that I'd suggest using Fedora; 29 is not so far from CentOS 8 and if you find something that needs fixing, at least it can be useful as we move forward.

[1] https://zuul.opendev.org/t/openstack/build/c44acb26555147d6ae1f7e19498ff5d1/log/job-output.txt#6267
[2] https://lists.rdoproject.org/pipermail/dev/2019-October/009155.html
[3] https://review.opendev.org/#/c/688614/

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0