Update patch set 6
Patch Set 6:

> Will ovn become the new default for neutron and replace
 > openvswitch?
 > 
 > There is also a neutron devstack plugin within the neutron repo,
 > maybe this code can (mostly) be hosted there?

Hi,

We are now starting process of moving ovn driver code to be in-tree neutron driver. So it will be the same like ovs and linuxbridge today. Because of that my idea was to move this code related to networking-ovn to devstack repo directly. As for today, if You want to deploy devstack with neutron with ovs or linuxbridge agent You don't need to do something like:

    enable_plugin neutron

so I don't think that this should be required if You will choose ovn as Q_AGENT (when this will be also in-tree).
What do You think Jens about this?

Patch-set: 6