Update patch set 6
Patch Set 6:

> I'm not sure really, adding >700 lines of code looks pretty
 > difficult to review, so if you could review this whole thing within
 > the neutron team, it would take some load off the only minimal
 > devstack core team.
 > 
 > I do also see your point about not needing the neutron devstack
 > plugin, but maybe the could happen as a second step once the code
 > is completely within neutron?
 > 
 > In case you decide that you still want to move into devstack
 > directly: Is this copied 1:1 from the neutron-ovn repo? I'd also
 > want to see some job(s) running and passing based on this before
 > approving.

It's not 1:1 because in networking ovn this functions are scattered around. Since we were going to move the code I decided to refactor it into one module with better comments and a clear interface (install_ovn, configure_ovn, init_ovn, start_ovn, stop_ovn and cleanup_ovn).

About having a job, yeah what I can do is create a job in Neutron tree with a depends-on on this patch. Would that be enough ?

Also, yeah, I feel the pain of having to review > 700 lines of code. Sorry for that. I can try to get some Neutron/networking-ovn folks to review it to help you out.

Patch-set: 6