Update patch set 6
Patch Set 6:

> Another thing: With this you are effectively un-deprecating the
 > neutron-legacy code, so you should actually make that explicit in a
 > patch that comes before this one.
 > 
 > Not sure about deprecating the new neutron code at the same time.

That's a huge chunk of work to be honest with you, apparently everybody still using and sourcing neutron-legacy.

If possible, we would like to have the networking-ovn code moved to the neutron tree at the very beginning of the development cycle to avoid having to freeze new features and bug fixes later on. So would be OK if I open a bug (high prio) to refactor this DevStack modules after this effort ?

Patch-set: 6