Glance support for Cinder store
Enable configuration of Cinder as a Glance
backend store.  This patch is mostly a POC
for now.  Need to determine what config options
etc. should be used to control this.

Set GLANCE_STORE=cinder and add glance_store
to LIBS_FROM_GIT to enable.

Moves the first cirros image upload to Glance
until after the Cinder service has started.

Finally, if we want to enable swift as glance store
we need to set $GLANCE_STORE=swift.

Change-Id: If979a76ec4e7a2649b2c1a003cb9cef236559db4