Update patch set 2
Patch Set 2: Code-Review-1

Is there a way to verify that this actually does what it claims?

I'm also worried that it may affect gate stability negatively, is anyone actually using this kind of setup in production? And if this makes for a better default experience, wouldn't it be better to make it the default in nova?

Patch-set: 2
Label: Code-Review=-1