Update patch set 3
Patch Set 3:

> > Is there a way to verify that this actually does what it claims?
 > >
 > > I'm also worried that it may affect gate stability negatively, is
 > > anyone actually using this kind of setup in production? And if
 > this
 > > makes for a better default experience, wouldn't it be better to
 > > make it the default in nova?
 > 
 > Hmm, that's a good point. Which tests are failing? I went looking
 > at the nova-compute logs for the Tempest job and couldn't see the
 > RNG device in the XML for the various instances, however, this
 > seems to be because we don't use the flavors created in 'lib/nova'
 > but rather those created in 'lib/tempest' (m1.nano, by default). Do
 > we need to modify that file too?

I've updated the change to include updates for the tempest flavors.

Testability was my concern too. I am definitely not the expert here was simply trying to get the ball rolling. I think questions of testing may be best deferred to nova?

As for checking if this improves gate stability I think this may be a case of merge and mark point in time against related e-r bugs. Then check again in a week and see if trends have changed. If they change for the better great. If there is no change or change for the worse we should consider reverting.

Patch-set: 3